kafka_consumer/requirements.txt
neo4j
kafka-python-ng